+++ api/src/main/scala/Routes.scala
@@ -1,17 +1,11 @@
 import akka.http.scaladsl.server.Directives._
 import api._
 
-trait Routes extends LecturesListApi
-            with LecturesDetailsApi
-            with LecturesCreateApi
-            with LecturesDeleteApi
-            with LecturesUpdateApi {
+trait Routes extends LecturesApi {
 
   val routes = pathPrefix("v1") {
     lecturesListRoute ~
-    lectureDetailsRoute ~
     lectureCreateRoute ~
-    lectureDeleteRoute ~
-    lectureUpdateRoute
+    lectureDetailsRoute
   } ~ path("")(getFromResource("public/index.html"))
 }
