--- api/src/main/scala/api/LecturesDeleteApi.scala
@@ -1,9 +1,0 @@
-package api
-
-import akka.http.scaladsl.server.Directives._
-
-trait LecturesDeleteApi {
-  val lectureDeleteRoute = (path("lectures"/ IntNumber ) & delete) { lectureId =>
-    complete("Trying to delete lecture with id: " + lectureId)
-  }
-}
