--- api/src/main/scala/api/LecturesDetailsApi.scala
@@ -1,10 +1,0 @@
-package api
-
-import akka.http.scaladsl.server.Directives._
-
-trait LecturesDetailsApi {
-  val lectureDetailsRoute =
-    (path("lectures"/ IntNumber ) & get) { lectureId =>
-      complete("LectureId: " + lectureId)
-    }
-}
