--- api/src/main/scala/service/LecturesCreateService.scala
@@ -1,12 +1,0 @@
-package service
-
-import dao.LecturesCreateDao
-import model.LectureCreate
-import reactivemongo.api.commands.WriteResult
-
-import scala.concurrent.Future
-
-object LecturesCreateService {
-  def create(lecture: LectureCreate): Future[WriteResult] =
-    LecturesCreateDao.create(lecture)
-}
