--- api/src/main/scala/service/LecturesDeleteService.scala
@@ -1,5 +1,0 @@
-package service
-
-class LecturesDeleteService {
-
-}
