--- api/src/main/scala/service/LecturesDetailsService.scala
@@ -1,5 +1,0 @@
-package service
-
-class LecturesDetailsService {
-
-}
