--- api/src/main/scala/service/LecturesUpdateService.scala
@@ -1,5 +1,0 @@
-package service
-
-class LecturesUpdateService {
-
-}
